refactor: extract FriendPicker from tag-people for reuse
